+++ src/open_api/responses/batches_response.rs
@@ -5,12 +5,12 @@ use poem_openapi::Object;
 
 #[derive(Clone, Debug, Object)]
 pub struct BatchesResponse {
-    total: i32,
+    total: i64,
     batches: Vec<Batch>,
 }
 
 impl BatchesResponse {
-    pub fn new(total: i32, block_batches: Vec<BlockBatch>) -> Self {
+    pub fn new(total: i64, block_batches: Vec<BlockBatch>) -> Self {
         let batches = block_batches.into_iter().map(Into::into).collect();
 
         Self { total, batches }
